Move log_request and logger to server-lib directory
