+++ web-dapp/server-lib/logger.js
@@ -1,12 +1,12 @@
 "use strict";
 
-const logDate = new Date().toISOString();
-
 module.exports = {
   log: (msg) => {
+    let logDate = new Date().toISOString();
     console.log(logDate, msg);
   },
   error: (msg) => {
+    let logDate = new Date().toISOString();
     console.error(logDate, msg);
   }
 };
