+++ web-dapp/server-lib/session-stores/redis.js
@@ -1,6 +1,6 @@
 'use strict';
 const config = require('../../server-config');
-const logger = require('../../utils/logger');
+const logger = require('../logger');
 const redis = require('redis'); // https://github.com/NodeRedis/node_redis
 const prelog = '[redis] ';
