+++ web-dapp/server-lib/session_store.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const config = require('../server-config');
-const logger = require('../utils/logger');
+const logger = require('./logger');
 
 const type = config.session_store.type;
 logger.log('[session_store] using session store of type: ' + type);
