+++ web-dapp/server.js
@@ -3,7 +3,7 @@
 const express = require("express");
 const body_parser = require("body-parser");
 const helmet = require("helmet");
-const logger = require("./utils/logger");
+const logger = require("./server-lib/logger");
 const config = require("./server-config");
 const recalc_price = require("./server-lib/recalc_price");
